ViewController done and propagated. Menu actions still tangled
Moving some new class to their respective package (using the same naming
convention as the scenebuilder project)
First draft for a lightweight docking api.
Introduced the concept of view/dock. A view requests to be docked into a
dock.
BuiltinLibrary and UserLibrary are now spring components.
Starting to untangle the project by deleting static getters in favor of
spring autowiring.